requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
wikipedia
pypdf
uvicorn
faiss-cpu
sse_starlette
fastapi
chromadb
llama-index
google-generativeai
llama-index-llms-gemini
IPython
llama-index-embeddings-gemini
langchain-groq
groq
cassio
beautifulsoup4